pagina crud final touches
pagina crud final touches

+++ app/Http/Controllers/PageController.php
@@ -12,7 +12,7 @@
 	
 	use App\Http\Requests;
 	use App\Http\Requests\Page\PageRequest;
-	
+	use Flash;
 	use Auth;
 	use Redirect;
 	use Request;
@@ -65,8 +65,8 @@
         public function store(PageRequest $request)
         {
             $introduction = $this->introrepo->create([
-            	'pageId' => 2,
             	'title' => $request->title, 
+            	'subtitle' => $request->subtitle,
             	'text' => $request->content,
             	]);
 
@@ -114,6 +114,11 @@
          */
         public function show($id)
         {
+            
+        	if($this->redirectHome($id)){
+        		return Redirect::route('home.index');
+        	} 
+
             $page = $this->pagerepo->get($id);
            	if(isset($page)){
 	            if($page->sidebar){
@@ -124,7 +129,7 @@
 	            }
 	        }else {
 	        	return view('errors.404');
-	        }
+	        } 
         }
 
         /**
@@ -135,6 +140,10 @@
          */
         public function edit($id)
         {
+            if($this->redirectHome($id)){
+        		return Redirect::route('home.index');
+        	}
+
             $page = $this->pagerepo->get($id);
             if(isset($page)){
             	return View('page.edit', compact('page'));
@@ -152,13 +161,17 @@
          */
         public function update($id, PageRequest $request)
         {
-            
+           if($this->redirectHome($id)){
+        		return Redirect::route('home.index');
+        	}
+
         	$old = $this->pagerepo->get($id)->sidebar;
         	$new = $request->sidebar;
 
         	// update introduction
         	$introduction = $this->introrepo->get($request->intro_id);
         	$introduction->title = $request->title;	
+        	$introduction->subtitle = $request->subtitle;
         	$introduction->text = $request->content;
 
         	$this->introrepo->update($introduction);
@@ -201,6 +214,12 @@
          */
         public function destroy($id)
         {
+
+        	if($this->redirectHome($id)){
+        		Flash::success('U kunt de homepagina niet verwijderen');
+        		return Redirect::route('home.index');
+        	}
+
         	$page = $this->pagerepo->get($id);
             if($page->sidebar){
   				$this->sidebarrepo->deleteAllFromPage($id);
@@ -238,4 +257,21 @@
 
         	}
         }
+
+        /* 
+         * redirectHome will redirect if the page is the homepage.
+         * The homepage has different edit functions and a different pageview.
+         * 
+         */
+
+
+        private function redirectHome($id){
+			if($id === '1'){
+				Flash::success('U bent succesvol naar de homepagina begeleid.');
+				return true;
+        	} else {
+        		return false;
+        	}
+        }
+
     }
